nano requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
jinja2
python-multipart
pandas
openpyxl
starlette
